try to improve this shit

+++ sem6/task1/dense_matrix.cpp
@@ -10,9 +10,9 @@ void dense_matrix::print() const {
     std::cout << "Matrix shape: " << nrows << " x " << ncols << std::endl;
     if (if_empty)
         return;
-    for (uint64_t i = 0; i < nrows; i++) {
+    for (uint32_t i = 0; i < nrows; i++) {
         std::cout << "row " << i << " : ";
-        for (uint64_t j = 0; j < ncols; j++)
+        for (uint32_t j = 0; j < ncols; j++)
             std::cout << val[i * nrows + j] << " ";
         std::cout << std::endl;
     }
@@ -26,8 +26,8 @@ void dense_matrix::generate() {
     std::normal_distribution<> dis(0, 1000);
 //    std::binomial_distribution<> dis(1000, 0.5);
     val.resize(0);
-    for (uint64_t i = 0; i < nrows; i++) {
-        for (uint64_t j = 0; j < ncols; j++) {
+    for (uint32_t i = 0; i < nrows; i++) {
+        for (uint32_t j = 0; j < ncols; j++) {
             val.push_back(dis(gen));
         }
     }
@@ -35,30 +35,30 @@ void dense_matrix::generate() {
     if_empty = false;
 }
 
-std::vector<double> dense_matrix::get_column(uint64_t index) const {
+std::vector<double> dense_matrix::get_column(uint32_t index) const {
     assert(index < ncols);
     std::vector<double>column;
-    for (uint64_t i = 0; i < nrows; i++)
+    for (uint32_t i = 0; i < nrows; i++)
         column.push_back(val[i * nrows + index]);
 
     return column;
 }
 
-std::vector<double> dense_matrix::get_column(uint64_t col, uint64_t row) const {
+std::vector<double> dense_matrix::get_column(uint32_t col, uint32_t row) const {
     assert(col < ncols);
     assert(row < nrows);
     std::vector<double> column;
-    for (uint64_t i = row; i < nrows; i++)
+    for (uint32_t i = row; i < nrows; i++)
         column.push_back(val[i * nrows + col]);
 
     return column;
 }
 
-std::vector<double> dense_matrix::get_minor_column(uint64_t index) const {
+std::vector<double> dense_matrix::get_minor_column(uint32_t index) const {
     assert(index < ncols);
     assert(index < nrows);
     std::vector<double> column;
-    for (uint64_t i = index; i < nrows; i++)
+    for (uint32_t i = index; i < nrows; i++)
         column.push_back(val[i * nrows + index]);
 
     return column;
