+++ sem6/task1/dense_matrix.h
@@ -9,20 +9,20 @@ public:
     std::vector<double> val;
 
     bool if_empty = true;
-    uint64_t nrows = 0, ncols = 0;
-    uint64_t nonzeros = 0;
+    uint32_t nrows = 0, ncols = 0;
+    uint32_t nonzeros = 0;
 
 //    ~dense_matrix;
 //    allocate
 
-    dense_matrix(uint64_t _nrows, uint64_t _ncols) : nrows(_nrows), ncols(_ncols) {
+    dense_matrix(uint32_t _nrows, uint32_t _ncols) : nrows(_nrows), ncols(_ncols) {
         nonzeros = nrows * ncols;
         val.resize(nonzeros);
     }
 
     void print() const;
     void generate();
-    std::vector<double> get_column(uint64_t ) const;
-    std::vector<double> get_column(uint64_t, uint64_t ) const;
-    std::vector<double> get_minor_column(uint64_t ) const;
+    std::vector<double> get_column(uint32_t ) const;
+    std::vector<double> get_column(uint32_t, uint32_t ) const;
+    std::vector<double> get_minor_column(uint32_t ) const;
 };
